requirements.txt
appdirs==1.4.3
backports.weakref==1.0rc1
bleach==1.5.0
funcsigs==1.0.2
html5lib==0.9999999
Markdown==2.6.8
MIDIUtil==1.1.3
mock==2.0.0
msgpack-python==0.4.8
numpy==1.13.0
packaging==16.8
pandas==0.20.1
pbr==3.1.1
pkg-resources==0.0.0
protobuf==3.3.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-midi==0.2.4
pytz==2017.2
scipy==0.19.0
six==1.10.0
tensorflow==1.2.1
Theano==0.9.0
theano-lstm==0.0.15
tqdm==4.14.0
Werkzeug==0.12.2
